+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Bump golangci-lint from 1.55.1 to 1.61.0 (developers should update to this version).
 * Update ctclient tool to support SCT extensions field by @liweitianux in https://github.com/google/certificate-transparency-go/pull/1645
 * Bump go to 1.23
+* [ct_hammer] support HTTPS and Bearer token for Authentication.
 
 ## v1.3.1